requirements.txt
streamlit==1.7.0
scikit-learn~=1.0.2
pandas~=1.4.1
numpy~=1.22.3
Pillow~=9.0.1
nltk~=3.7
geopandas~=0.10.2
langdetect~=1.0.9
wordcloud~=1.8.1
yellowbrick~=1.4
folium~=0.12.1
streamlit-folium==0.6.2